@pcrespov review: moved project fixture to package
